Make NXT Bluetooth connection manager listen for connections.
The NXT bluetooth stack can only have one incoming connection at a
time (i.e. it can only be a slave in a single piconet), so after
accepting a connection the manager won't try to accept another until
that one has closed.

This patch also corrects a logic bug in the connection listener and
adds a loop to ensure that enough bytes are read when receiving the
remote address on connection startup.